requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# AI & LLM
anthropic==0.18.0
langchain==0.1.10
langchain-anthropic==0.1.4
langgraph==0.0.25
llama-index==0.10.12

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
pydantic==2.6.0

# Visualization
plotly==5.18.0
folium==0.15.1
matplotlib==3.8.2

# Database & SQL
sqlalchemy==2.0.25
duckdb==0.9.2

# ML Tracking (optional for advanced tracing)
mlflow==2.10.0

# Utilities
requests==2.31.0
tqdm==4.66.1

httpx<0.27